# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Net Zone CRM API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /customers:
    get:
      operationId: listCustomers
      tags: [customers]
      summary: List all customers
      security:
        - bearerAuth: []
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - name: category
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of customers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Customer"
    post:
      operationId: createCustomer
      tags: [customers]
      summary: Create a new customer
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CustomerInput"
      responses:
        "201":
          description: Customer created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Customer"
...
  /calls:
    get:
      operationId: listCalls
      tags: [calls]
      summary: List call records
      security:
        - bearerAuth: []
      parameters:
        - name: period
          in: query
          schema:
            type: string
            enum: [today, yesterday, week]
        - name: agentId
          in: query
          schema:
            type: string
        - name: type
          in: query
          schema:
            type: string
        - name: customerId
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of calls
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Call"
    post:
      operationId: createCall
      tags: [calls]
      summary: Create a call record
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CallInput"
      responses:
        "201":
          description: Call created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Call"
...
  /reminders:
    get:
      operationId: listReminders
      tags: [reminders]
      summary: List reminders
      security:
        - bearerAuth: []
      parameters:
        - name: customerId
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of reminders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Reminder"
    post:
      operationId: createReminder
      tags: [reminders]
      summary: Create a reminder
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReminderInput"
      responses:
        "201":
          description: Reminder created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Reminder"
  /reminders/{id}:
    put:
      operationId: updateReminder
      tags: [reminders]
      summary: Update reminder status
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReminderUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Reminder"
    delete:
      operationId: deleteReminder
      tags: [reminders]
      summary: Delete reminder
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Deleted

  # ── CATEGORIES ──────────────────────────────────────────────────────────────
...
  /reports/calls:
    get:
      operationId: getCallsReport
      tags: [reports]
      summary: Calls report with date and agent filters
      security:
        - bearerAuth: []
      parameters:
        - name: from
          in: query
          schema:
            type: string
        - name: to
          in: query
          schema:
            type: string
        - name: agentId
          in: query
          schema:
            type: string
        - name: type
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Calls report
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Call"
  /reports/summary:
    get:
      operationId: getReportSummary
      tags: [reports]
      summary: Summary counts grouped by day and type
      security:
        - bearerAuth: []
      parameters:
        - name: from
          in: query
          schema:
            type: string
        - name: to
          in: query
          schema:
            type: string
        - name: agentId
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Summary report
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReportSummary"
  /reports/customers:
    get:
      operationId: getCustomersReport
      tags: [reports]
      summary: Customers report with call counts
      security:
        - bearerAuth: []
      parameters:
        - name: category
          in: query
          schema:
            type: string
        - name: from
          in: query
          schema:
            type: string
        - name: to
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Customers with call stats
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Customer"

  # ── AGENTS ──────────────────────────────────────────────────────────────────
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Customer:
      type: object
      required: [id, name, mobile, email, address, category, createdAt, updatedAt]
      properties:
        id:
          type: string
        name:
          type: string
        mobile:
          type: string
        alternateNumber:
          type: ["string", "null"]
        email:
          type: string
        company:
          type: ["string", "null"]
        address:
          type: string
        city:
          type: ["string", "null"]
        category:
          type: string
        notes:
          type: ["string", "null"]
        followUpDate:
          type: ["string", "null"]
        totalCalls:
          type: ["integer", "null"]
        lastCallAt:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
    CustomerInput:
      type: object
      required: [name, mobile]
      properties:
        name:
          type: string
        mobile:
          type: string
        alternateNumber:
          type: string
        email:
          type: string
        company:
          type: string
        address:
          type: string
        city:
          type: string
        category:
          type: string
        notes:
          type: string
        followUpDate:
          type: string
...
    Call:
      type: object
      required: [id, customerId, agentId, type, duration, createdAt]
      properties:
        id:
          type: string
        customerId:
          type: string
        customerName:
          type: ["string", "null"]
        agentId:
          type: string
        agentName:
          type: ["string", "null"]
        type:
          type: string
          enum: [incoming, outgoing, missed]
        duration:
          type: integer
        phoneNumber:
          type: ["string", "null"]
        category:
          type: ["string", "null"]
        remark:
          type: ["string", "null"]
        followUpDate:
          type: ["string", "null"]
        reminderDate:
          type: ["string", "null"]
        deviceId:
          type: ["string", "null"]
        createdAt:
          type: string
    CallInput:
      type: object
      required: [customerId, type, duration]
      properties:
        customerId:
          type: string
        type:
          type: string
          enum: [incoming, outgoing, missed]
        duration:
          type: integer
        phoneNumber:
          type: string
        category:
          type: string
        remark:
          type: string
        followUpDate:
          type: string
        reminderDate:
          type: string
        deviceId:
          type: string
...
    Reminder:
      type: object
      required: [id, customerId, agentId, dateTime, title, status, priority, createdAt]
      properties:
        id:
          type: string
        customerId:
          type: string
        agentId:
          type: string
        title:
          type: string
        dateTime:
          type: string
        notes:
          type: ["string", "null"]
        status:
          type: string
          enum: [pending, completed, overdue]
        priority:
          type: string
          enum: [low, medium, high]
        createdAt:
          type: string
    ReminderInput:
      type: object
      required: [customerId, title, dateTime, priority]
      properties:
        customerId:
          type: string
        title:
          type: string
        dateTime:
          type: string
        notes:
          type: string
        priority:
          type: string
          enum: [low, medium, high]
    ReminderUpdate:
      type: object
      properties:
        status:
          type: string
          enum: [pending, completed, overdue]
        title:
          type: string
        dateTime:
          type: string
        notes:
          type: string

    # ── CATEGORIES ────────────────────────────────────────────────────────────
...
    ReportSummary:
      type: object
      required: [totalCalls, incomingCalls, outgoingCalls, missedCalls, avgDuration, byDay]
      properties:
        totalCalls:
          type: integer
        incomingCalls:
          type: integer
        outgoingCalls:
          type: integer
        missedCalls:
          type: integer
        avgDuration:
          type: number
        byDay:
          type: array
          items:
            $ref: "#/components/schemas/DailyCallStat"
    DailyCallStat:
      type: object
      required: [date, total, incoming, outgoing, missed]
      properties:
        date:
          type: string
        total:
          type: integer
        incoming:
          type: integer
        outgoing:
          type: integer
        missed:
          type: integer